cleaned up the code and orginized it
