+++ src/userDashboardTimer.js
@@ -1,16 +1,19 @@
-// importing
-import { accountLogout } from "./userDashboardLogic.js";
-
 // selecting elements
 const dashboardBottomText1 = document.getElementsByClassName(
   "dash-board-Bottom-Text-1"
 )[0];
+// logout btn
+const dashboardLogoutBtn = document.getElementsByClassName(
+  "dashboard-Logout-Btn"
+)[0];
+
+let minCount = 0;
+let seconds = 0;
 
 // our functions
 export const userDashBoardTimer = () => {
-  let minCount = 10;
-  let seconds = 0;
-
+  minCount = 10;
+  seconds = 0;
   let timerInterval = setInterval(() => {
     if (minCount > 0) {
       if (seconds > 0) {
@@ -26,5 +29,29 @@ export const userDashBoardTimer = () => {
       accountLogout();
     }
     dashboardBottomText1.textContent = `You will be logged out in 0${minCount}:${seconds}`;
-  }, 1000);
+  }, 50);
 };
+
+// logout functions
+let accountLogout = () => {
+  dashboardPageEl.classList.remove("active");
+  loading.classList.remove("active");
+  let loginCounter = 0;
+  clearInterval(timerInterval);
+  minCount = 0;
+  seconds = 0;
+
+  let logininterval = setInterval(() => {
+    if (loginCounter < 10) {
+      loginCounter++;
+    } else {
+      loginCounter = 0;
+      clearInterval(logininterval);
+      landingPageEl.classList.remove("active");
+      loading.classList.add("active");
+    }
+  }, 500);
+};
+
+// event linsters
+dashboardLogoutBtn.addEventListener("click", accountLogout);
